Added specification for new features.
Use manually set uid values in named attribute, 
validating that they have not already been used, 
and in the range of values specified.

Allow for the scope of the elements considered 
when building a header file to be altered:

-Global: 	all elements of a given type.
-Scoped: 	all elements of a given type, below an
			element of a given type (the processor
			instruction would be at the ancestor's
			type node in the schema, naming the 
			descendant type to pool as a parameter
			).

Element naming collisions must be flagged, once 
any hierarchy stripping has been applied.